+++ NEWS.md
@@ -1,13 +1,22 @@
 # vapour 0.4.0
 
+## BREAKING CHANGES
+
+* `vapour_read_raster()` now returns a list of (one) numeric vector. This change is so 
+ a future version can return more than one band for this function and for `vapour_warp_raster()`. 
+ 
+## NEW FEATURES
+
 * New data set `tas_wkt` a Well-Known-Text string to simplify an example
  in the documentation. 
+ 
 * The function `vapour_read_raster()` now returns a list with a single vector of
  values from the requested band. In future this may return values from multiple
  bands. The new function `vapour_warp_raster()` uses the same scheme.
 
 * New function `vapour_warp_raster()` to return a warped version from
  a raster source. 
+ 
 * Fixed static linking for MacOS build, thanks(!!) to Jeroen Ooms for guidance 
  and thanks to CRAN for reporting the issue.